+++ CHANGELOG.md
@@ -13,7 +13,9 @@
    - Estimate bridge users by country based on requests by country, if
      available, to get more accurate numbers than those obtained from
      unique IP address counts.
-   - Update to metrics-lib 2.11.0 and ExoneraTor 4.4.0.
+   - Update to metrics-lib 2.12.2 and ExoneraTor 4.4.0.
+   - Switch from processing Torperf .tpf to OnionPerf analysis .json
+     files.
 
  * Minor changes
    - Make Jetty host configurable.